Remove the old "edit tags" dropdown.
It's redundant with the recent searches display in the regular tag dropdown, and it's
ancient and not very useful.  Remove it, since it's a pain to keep working.

Make the edit button that used to toggle to this view instead toggle editing in the
dropdown, replacing the big floating edit button.