Widget visibility stuff.
- Change visible-widget to hidden-widget, so if we're using something else for visibility, we
don't have to set visible-widget to keep a widget visible.  This also removes one bit of class
list noise from widgets, since most widgets are visible.
- Add visible_recursively, which is true if a widget and its ancestors are all visible.
- Add on_visible_recursively_changed, which is called when visible_recursively changes.

This removes the need for view_hidden_listener, which was an awkward use of events, and makes
whether a widget is actually visible easier to access.