More dropdown refactoring.
Use dropdown_box_opener for dropdowns in the context menu.  This makes them behave like other
popups: the menu will be created on demand instead of in advance, pushed left to fit if needed
so they're more likely to fit on screen, etc.

Move the hover handling for the desktop image UI into image_ui (needed for changes to bookmark
dropdown handling).

Add ppixiv.bookmark_tag_dropdown_opener to handle opening dropdowns.

Changed bookmark_button_widget onedited to an event, so it's easier for the tag dropdown
handling to listen to it.