More dropdown cleanup.
Use dropdown_box_opener for the follow dropdown.  (I think this is the last one...)

Made scrollers for the dropdowns work more consistently.

Remove visibility handling from follow_widget, since dropdown_box_opener handles this.

If a dropdown is on the bottom half of the screen, put it above the button instead of
below.  This fixes tiny bookmark dropdowns when clicking on a thumbnail near the bottom
of the screen.

Fix local_search_dropdown_widget having an invalid width if tag-dropdown-width is invalid.

dropdown_box_opener now handles all dropdown positioning, so moved helpers.set_max_height
into it.