Improve startup time on mobile.
Use the initialization info on mobile, and request the mute info separately.
Cache mutes when possible so we know them as well as possible until they're
loaded.  This speeds up initialization, and helps prevent having to go async
before setting up, which makes it more likely that we can shut down stuff on
the page.