Improve hiding the document while we set up.
Before, we set document.hidden to true to hide the page until we had something to display.  Instead,
set the background to #000 and set a brightness filter to 0.  This has the same effect, but blanks
the screen to black instead of white, so there's less of a flashbang on load.  There's still some
depending on the browser, since we aren't run immediately.