Mobile dragging adjustments.
- Give TouchScroller explicit state callbacks to match WidgetDragger.  It would be nice to
avoid duplicating these, but I'm not sure if it's worth trying to share code between these.
- In image_viewer_mobile, cancel animations when TouchScroller goes active instead of immediately
on press, which prevents them from being stopped if we end up going into an image change drag.
That allows the animation to continue, and avoids strange-looking behavior as the post-animation
zoom snapping happens at the same time as the viewer panning off-screen, which often makes it look
like the image is jumping downwards slightly.  This also means we don't stop the animation if
the menu is opened by dragging down.
- Don't start dragging on pointerup unless we were dragging, not if we see pointer releases
when we're idle.  TouchScroller will see the touch release when we didn't activate it and didn't
stop the animation and it shouldn't start_fling in that case.