Remove current_url in set_item.
I'm not sure what this was for since I never bothered to comment it, and it doesn't seem to
make sense.  Data source UI URLs should be relative to the data source URL, not the current
URL, since a data source can be used independantly of the active navigation and the current
URL might have things data source UI links shouldn't keep, like the media ID being viewed,
which can cause weird navigations.