Adjust thumbnail creation.
If we have no nearby thumb at all, it's the same as reaching the end of the list, so extend start_idx
and end_idx.

This fixes only creating one or two thumbs on the first refresh when forced_media_id is specified.
We'd refresh again almost immediately once that thumb became nearby, but this caused transitions
from screen_illust to go to the wrong place since it went where the thumbnail was after the first
refresh.