Refactor the top-level app and startup.
Factor out AppStartup.  This centralizes the stuff that happens before the app
actually starts: figuring out if we want to start, setting up the launch button,
cleanup_environment, etc.  This class shouldn't require anything outside of itself,
so it can be run before we have access to anything else.  This is meant to make
it easier to do this stuff without actually loading the rest.  It should also
eventually allow moving the rest of the code into modules.  This class can't be
in a module, because loading a module is always async and this setup needs to happen
without going async, so it happens before Pixiv scripts run.

AppStartup takes over chunks of page_manager, early init in main_controller, and
helpers.cleanup_environment.  The rest of main_controller is now ppixiv.App.

Functions like helpers.fetch are now window.realFetch, so this can be set up
without having access to helpers to store them on.