Remove the second preview image in ViewerUgoira.
It seems to cause Firefox to flash white as it switches.  It doesn't seem worth tracking
this down, since this probably isn't worth doing anyway as we should get the first animation
frame just as quickly as we'd download the static image.  This also makes this more similar
to the video viewer.