Rework the user links into the avatar dropdown.
The user link strip was a mass of icons that had to be hovered to see what they
were.  Sometimes what you wanted was obvious, but it often just had a bunch of
nameless shopping carts and link icons.  It could get huge for users that have
lots of links in their profile, so we had to limit the number of links that could
be displayed, and it competed with the user's name in the UI for space.  It also
didn't work at all on mobile, where you can't hover to find out what something is.

This replaces that, putting that in the avatar dropdown with follow/unfollow.

- It's a lot easier to see what things are, since it's just a regular menu.
- It can be accessed from the popup menu and the desktop hover UI.  It was only available
from the artist view before.
- It works on mobile, so we don't need to figure out a completely separate UI there.

Follow tags are in a submenu.  This makes things make more sense if you have a lot
of follow tags, and avoids the dropdown shifting around as much as it loads.  (It still
does a bit for Booth and Fanbox links due to API limitations.)

This uses fewer site-specific icons, preferring to use generic "shopping cart", etc. icons.
The specific icons aren't necessary since they have labels, and the standard icons make it
easier to skim the list.