When navigating from an image to an artist, scroll to the top instead of to the image we came from.
Going to the artist for a post is usually discovery, where you want to see what else
the user has posted, and scrolling to the image we came from isn't really wanted.  This
also makes more sense if we show artist info at the top of the search scroller (the
mobile UI may do this).