Refactor ViewerImages to use the aspect ratio instead of the image resolution.
All we actually care about is the aspect ratio, so this lets us display the image without
having to wait for media info as long as we've previously loaded the page's thumbnail.  Since
most navigations come from searches, this happens most of the time, and fixes the delay when
clicking a manga page in search results.

Some things still require the real image size (cropping and "actual" zoom mode), so those
will still cause the display to be delayed.