Refactor the /ids mode of DataSource_VView.
The /ids request returns all files in a directory, so when viewing a very large directory
we can request image data as it's viewed instead of having to do it all upfront.  It relies
on SearchView being able to handle being given a huge list of IDs and requesting data as
needed.  It also allows jumping to an arbitrary file in a huge directory, since we have the
full list.

It's a bit of a pain, since it's different from all other data sources and it makes it hard
to adjust SearchView.  Everything else is page-based and SearchView can just load a page at
a time.

Make this work more like a paginated view by looking for the file we started on and making
its page the initial page.