Rework SearchView.
- Replace the flex-wrap box with separate flex rows.  Flexbox isn't good enough to
do this on its own and it was causing too many problems, and this gives a lot more
control.
- We only add thumbs to the beginning or end of the list.  If we need to add thumbs
in the middle, or if the window size or settings change, clear the list and start over.
- Once we fill a row of thumbs, it's finalized and never changed.  If we add a row to
the beginning and there's no active row, we'll create a new row for it with nothing
else on it.  This creates orphans where it's possible to have just one thumb at the
top of the screen if you started in the middle and scrolled up, but that's not too
common and this lets us add to the beginning of the list freely without the rest of
the screen shifting around.
- Loading earlier pages now works globally.  The "load previous page" button has
been removed.
- We can now cleanly handle non-square thumbs.  This was hard before because adding
thumbs at the beginning would cause everything to shift around (with square thumbs
we just added a whole row at a time), but doing things per-row fixes this limitation.
- Non-square thumbs are sized in a way that still fills the screen horizontally, and
tries to minimize ugly wasted space caused by interleaving portrait and landscape
thumbnails.