Update the thumbnail sizing algorithm.
Previously, we clamped the column count, and looked for a column count that gave
thumbs close to a size set by the thumbnail size setting.  This doesn't make sense
with aspect ratio thumbs, and it limits thumbs more than it needs to.

Instead, clamp the approximate total number of thumbs we'll show at once.  This
makes more sense for different window sizes: if you set thumbs to small and have
a wide window that isn't very tall, we'll allow lots of columns to fill it in,
instead of clamping at 5.

This removes the need for a separate value for mobile, since it generalizes to
mobile screen sizes better.