Rework safe partial media info.
Previously, if non-full media info was requested but we actually had full info, we'd
create a new mediaInfo with just the partial keys, so we don't accidentally use keys
that might not be available.  With MediaInfo this is a pain: making a copy means we
can't modify the copy, since changes won't be seen on the original.

Instead, wrap MediaInfo objects with a proxy that checks the keys being accessed.  This
way they're still really using the same underlying object.