Add a leading slash to imports.
This allows the script directory to be added with a single import map entry for /vview.
We don't do this for a couple reasons (that doesn't work when loading from blobs, which
we need to do when running as a user script, and relative paths aren't implemented in
ModuleImporter_Compat), but it's convenient to allow other code to load our scripts.