Support navigating to series from images on mobile.
This removes the "return to manga view" logic.  It tried to make it so clicking
the manga button would go back instead of loading a new search if the previous
state was the manga view.  It only works on mobile since the browser history API
is broken and doesn't let you see your own history states, and normally you exit images
on mobile by swiping up anyway, not by using the menu.  It's also hard to make
it work for series, where we have to load illust info to know if there's a series
to go to.