Rework loading using esbuild.
Remove ModuleLoaders.  ModuleLoaders_Compat was an ugly mess to allow us to use JS modules
from within a user script.  (ModuleLoaders_Native didn't actually work due to design flaws
in import maps when scripts are loaded from blobs.)  Instead, bundle the app using esbuild.
It handles things like packaging our resources in roughly the same way we were before.

This is also used when running natively with vview, instead of loading files directly.
It's hard to get esbuild to handle cache busting without using its API, which would require
pulling in a whole node installation.  The resources aren't big enough for this to be a
problem, and and it's probably better to just eliminate this difference between vview and ppixiv.

We generate the bundle each time it's requested.  esbuild is quick, and it's skipped if
the browser has the bundle cached.

The bundle is still packaged into a JS string for the user script, since we still need to run
it in the page context and not the userscript context.

This should remove the weird restrictions on JS import syntax, and allow using resource
imports.