Replace the manga progress bar.
The display in the hover was awkward, since you can't always mousewheel to scroll
to the next image while hovering over it (often it'll scroll its contents instead).
Replace it with an optional progress bar at the bottom of the image that flashes
briefly when changing pages.  This is disabled by default.