Rework view-in-explorer.
Switch to using a "vview" scheme with the command ("view-in-explorer") in the host.  Putting
the path there is annoying because it always wants to have a leading slash, but we have both
Windows absolute paths that don't and UNC paths that do.  This also allows adding more handlers
without needing to change the scheme registration.